add missing status in book_schemas.py
